keep the notification hidden when reconnecting to the VPN.
This will happen whenever the user changes connection status (wifi to data, etc). While it is useful for debugging, it might be confusing for the user, and they may not understand why the app is connecting and disconnecting frequently. It only ever takes a few seconds, but keeping that away from the user is probably a better idea